reorganize coms dir to seperate interests
