+++ radioapi/radioapi/coms/errors/__init__.py
@@ -1,0 +1,9 @@
+from .errors import (
+    ComsMessageParseError,
+    ComsDriverReadError,
+)
+
+__all__ = [
+    "ComsMessageParseError",
+    "ComsDriverReadError",
+]
