+++ radioapi/radioapi/coms/errors/errors.py
@@ -1,0 +1,5 @@
+class ComsMessageParseError(ValueError):
+    """Failed to parse a com message"""
+
+class ComsDriverReadError(Exception):
+    """Called when a ComsDriver cannot read a ComsMessage"""
