+++ radioapi/radioapi/coms/subscribers/__init__.py
@@ -1,0 +1,9 @@
+from .subscription import (
+    ComsSubscription,
+    OneTimeComsSubscription,
+)
+
+__all__ = [
+    "ComsSubscription",
+    "OneTimeComsSubscription",
+]
